lab2/pyproject.toml
[project]
name = "lab2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.12",
    "httpx==0.28.1",
    "pytest==8.3.5",
    "python-jose[cryptography]==3.4.0",
    "pyyaml==6.0.2",
    "uvicorn[standard]==0.34.0",
]
